// lib/classes/composition.interfaces.ts
export interface CompositionConstructor {
  canvasId: string;
  imgQuerySelector: string;
  options?: CompositionOpitons;
}

export interface CompositionOpitons {
  width: number;
  height: number;
}

export enum Channel {
  Red = "red",
  Green = "green",
  Blue = "blue",
  Alpha = "alpha",
}

export enum BlendMod {
  normal = "normal",
  add = "add",
}